file/fix: moved promo check to server action
