+++ database/drizzle/queries.ts
@@ -244,46 +244,3 @@ export const isExistingEmailAddress = async (
   });
   return !!exisitingEmail;
 };
-
-export const getPromoCodeStatus = async (
-  code: string
-): Promise<PromoCodeValidation> => {
-  try {
-    const promoCode = await db.query.promo_codes.findFirst({
-      where: eq(promo_codes.code, code),
-    });
-
-    if (!promoCode) {
-      return {
-        status: 'error',
-        message: 'invalid promo code.',
-      };
-    }
-
-    if (!promoCode.is_active) {
-      return {
-        status: 'error',
-        message: 'no longer active.',
-      };
-    }
-
-    if (promoCode.usage_count >= promoCode.max_usage) {
-      return {
-        status: 'error',
-        message: 'max uses reached.',
-      };
-    }
-
-    return {
-      status: 'success',
-      message: 'valid promo code.',
-      promoCode: promoCode,
-    };
-  } catch (error) {
-    console.error('Error validating promo code:', error);
-    return {
-      status: 'error',
-      message: `Error validating promo code: ${error}`,
-    };
-  }
-};
